requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
transformers==4.51.3
datasets==3.5.0
accelerate==1.6.0
peft==0.15.2
trl==0.17.0
torch>=2.2.0
bitsandbytes==0.45.5
pandas==2.2.3
sentence-transformers==4.1.0
faiss-cpu==1.11.0
numpy==2.2.5
